fix(extensions): move models to extensions/models.py; fix migration server_default
- Move ExtensionSettings and ExtensionEnabled out of superset/models/core.py
  into superset/extensions/models.py — core.py is the legacy mixed-bag the
  project has been breaking up; extension-scoped models belong next to settings.py.
- Update settings.py import to match new location.
- Fix migration server_default=\"1\" → sa.true() to avoid PostgreSQL warnings;
  aligns with sa.false()/sa.true() convention used throughout the migrations tree.
- Update migration docstring to mention both tables created.
